Trying to fix the CI
Deleting the lock file didn't solve anything. I think I found a way to specify to the Workflow to use the latest stable rust release, so we will give that a go.

I've also started making some tweaks to the icons in the UI, with hover affects, and adding the TouchArea that will be needed for unlocking the files.