Minimum working version
This I would say is a minimum working version. It pretty much does what I want. I may not make a lot of other changes.
I will cut a release and if I get not other suggestions, and don't find any bugs I'll probably bump this to 1.0

Updated release workflow to build for mac, and arm linux

Added in an app icon (although it's crude, might want to get a better one)

Added the ability add files and update the model accordingly