Some preliminary work for better error notification
I need to fix the issue where it will ignore an MER older than 5 because it checks for the lock file in the MER first but 4 and older don't have it. In the meantime I started some prep work for better error notification to the user when something goes wrong.