Cleaning up file handling and relaying messages to the user
I added in some better handling of older files. Now instead of failing, it will show the version for v4 and older files. It will also populate a note with information on the older versions. V4 files can be unlocked and the program allows this, but I need to prevent it from allowing a V3 file to try and be unlocked as this currently doesn't function.

I also want to add a bit of a progress bar to the slide over as the timer ticks down.

I've also upgraded to the latest version of Slint.